complete base dokeos addUser
git-svn-id: http://192.168.1.235/svn/repo/octopus@3274 7df5b637-c7b4-2643-9bff-991d5fa3ecdc

+++ src/com/lxitedu/framework/dao/SuperDAO.java
@@ -221,5 +221,5 @@ public abstract class SuperDAO {
       return (new Integer(result));
     }
   }
-  protected abstract Integer getNextPrimaryID() ;
+  protected abstract Long getNextPrimaryID() ;
 }
